More secure session handling (fixed)
- If an existing session isn't found in the session store, assign a new
session ID instead of reusing the given ID. This could prevent a session
fixation attack.
- If the session responds to #destroy use that method rather than #clear.
Thanks to hwilbanks@ranchopt.com for pointing this out.

@halloffame or @rift137 could I get this reviewed and merged quickly?
Once it is I'll roll out a new gem version.

Note: This is the second PR that removes some of the stuff that accidentally
got included in the original PR.